added very rough shop UI. Added ability to purchase weapons. working out a bug with incorrect abilities being shown

+++ Source/BeersAndBrawls/Items/UItem.h
@@ -31,7 +31,7 @@ public:
 	EItemType ItemType;
 
 	UPROPERTY(EditDefaultsOnly, BlueprintReadOnly)
-	float Cost;
+	int Cost;
 
 	UPROPERTY(EditDefaultsOnly, BlueprintReadOnly)
 	TArray<UAbilityInfo*> Abilities;
